HelpersTask789: Move purification to a new script
Pre-commit checks:
All checks passed ✅
