+++ dev_scripts_helpers/test/test_run_notebook.py
@@ -12,6 +12,7 @@ import helpers.hio as hio
 import helpers.hprint as hprint
 import helpers.hsystem as hsystem
 import helpers.hunit_test as hunitest
+import helpers.hunit_test_purification as huntepur
 
 _LOG = logging.getLogger(__name__)
 
@@ -151,7 +152,8 @@ def _compare_dir_signature(self: Any, dir_name: str, expected: str) -> None:
     # Remove references like:
     # $GIT_ROOT/core/dataflow/backtest/test/TestRunExperiment1.test3/tmp.scratch
     actual = actual.replace(dir_name, "$SCRATCH_SPACE")
-    actual = hunitest.purify_txt_from_client(actual)
+    text_purifier = huntepur.TextPurifier()
+    actual = text_purifier.purify_txt_from_client(actual)
     # Remove lines like:
     # $GIT_ROOT/core/dataflow_model/.../log.20210705_100612.txt
     actual = hunitest.filter_text(r"^.*/log\.\S+\.txt$", actual)
@@ -180,9 +182,7 @@ def run_cmd_line(
         "expected_pass=%s abort_on_error=%s", expected_pass, abort_on_error
     )
     _LOG.debug("cmd=%s", cmd)
-    rc = hsystem.system(
-        cmd, abort_on_error=abort_on_error, suppress_output=False
-    )
+    rc = hsystem.system(cmd, abort_on_error=abort_on_error, suppress_output=False)
     if expected_pass:
         self.assertEqual(rc, 0)
     else:
@@ -385,6 +385,7 @@ class TestRunNotebook2(hunitest.TestCase):
 @pytest.mark.superslow("~35 sec.")
 @pytest.mark.skip(reason="TODO(Juraj): HelpersTask21.")
 class TestRunNotebook3(hunitest.TestCase):
+
     def helper(
         self,
         fail: bool,
