+++ import_check/test/test_show_imports.py
@@ -6,13 +6,20 @@ import pytest
 
 import helpers.hio as hio
 import helpers.hunit_test as hunitest
+import helpers.hunit_test_purification as huntepur
 import import_check.show_imports as ichshimp
 
 _LOG = logging.getLogger(__name__)
 
 
+# #############################################################################
+# Test_show_imports
+# #############################################################################
+
+
 @pytest.mark.slow()
 class Test_show_imports(hunitest.TestCase):
+
     def create_io_dirs(self) -> None:
         dir_name = self.get_input_dir()
         hio.create_dir(dir_name, incremental=True)
@@ -83,7 +90,8 @@ class Test_show_imports(hunitest.TestCase):
             script_output = hio.from_file(script_output_filename)
             # Transform the output from the script by removing the dependencies
             # from the client.
-            purified_script_output = hunitest.purify_txt_from_client(
+            text_purifier = huntepur.TextPurifier()
+            purified_script_output = text_purifier.purify_txt_from_client(
                 script_output
             )
             purified_script_output = purified_script_output.replace(
